fix applet install error
The resource space allocated in the install() stage exceeds the limit of smart cards and cause the installation to fail.
Initialize the object after the applet is successfully installed consumes less resources and is compatible with more smartcards.
The smart card I am using is eSecuCard-S300 with Tongxin Microelectronics THD89 emv chip